Rename src/dymoprint to src/labelle with VS Code
