+++ src/labelle/barcode_writer.py
