+++ src/labelle/resources/fonts/LICENSE
