+++ src/labelle/resources/icons/barcode_icon.png
